% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VLabbott.R
\name{VLabbott}
\alias{VLabbott}
\title{Cleans Abbott files}
\usage{
VLabbott(return = FALSE, download = FALSE, data)
}
\arguments{
\item{return}{Specify whether to return a dataset. Valid arguments are 'errors' and 'percent'. Defaults to FALSE}

\item{download}{Specify whether to download fresh files from Google Drive. Defaults to FALSE}

\item{data}{Specify where to download files to or where they are stored}
}
\description{
Reads Abbott files, cleans data, and uploads them to Google Drive. Optionally returns one of two processed datasets.
}
