% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VLdownload.R
\name{VLdownload}
\alias{VLdownload}
\title{Pull Files from Database}
\usage{
VLdownload(dsn_database, dsn_hostname, dsn_port, dsn_uid, dsn_pwd, downpath)
}
\arguments{
\item{dsn_database}{Specify the name of your database.}

\item{dsn_hostname}{Specify the hostname}

\item{dsn_port}{Specify your port number as a character. e.g. 5432}

\item{dsn_uid}{Specify your username. e.g. "admin"}

\item{dsn_pwd}{Specify your password}

\item{downpath}{Specify the path to save downloaded files in}
}
\description{
This function pulls in all relevant files from the database and stores them locally. It only works with an approved IP address.
}
