% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VLroche.R
\name{VLroche}
\alias{VLroche}
\title{Cleans Roche files}
\usage{
VLroche(return = FALSE, download = FALSE, data)
}
\arguments{
\item{return}{Specify whether to return a dataset. Defaults to FALSE}

\item{download}{Specify whether to download fresh files from Google Drive. Defaults to FALSE}

\item{data}{Specify where to download files to or where they are stored}
}
\description{
Reads Roche files, cleans data, and uploads them to Google Drive. Optionally returns processed dataset.
}
