% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VLupload.R
\name{VLupload}
\alias{VLupload}
\title{Writes files to Google Drive}
\usage{
VLupload(downpath)
}
\arguments{
\item{downpath}{Specify the path where downloaded files are saved}
}
\description{
This function pulls in all relevant files from the database and stores them locally. It only works with an approved IP address.
}
